% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{AnnualInsolation}
\alias{AnnualInsolation}
\title{Annual insolation for any year of the last 5Mio years}
\usage{
AnnualInsolation(kyear, lat)
}
\arguments{
\item{kyear}{Thousands of years before present (0 to 5000).}

\item{lat}{Latitude in degrees (-90 to 90).}
}
\value{
daily average solar radiation in W/m^2 averaged over the year
}
\description{
Annual insolation for any year of the last 5Mio years
}
\details{
for details see daily_insolation_param
}
\examples{
plot(0:1000, AnnualInsolation(0:1000, 40), main = "mean annual insolation at 40degN",
xlab = "kyr BP", ylab = "W/m2", type = "l")
}
\author{
Thomas Laepple
}
