% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{calcE}
\alias{calcE}
\title{Calculate the Eccentricity anomaly}
\usage{
calcE(M, e, tol = 1e-08)
}
\arguments{
\item{M}{mean anomaly}

\item{tol}{tolerance}
}
\value{
Eccentricity anomaly E
}
\description{
Calculate the Eccentricity anomaly E from the mean anomaly M, iteratively
inversion of Keplers formula: E-e*sin(E)=M
based on Grossman, N.: Orbits under the Inverse Square Law, in
The Sheer Joy of Celestial Mechanics, edited by N. Grossman, pp. 41–61,
Birkhäuser Boston, Boston, MA., 1996.
}
\author{
Thomas Laepple
}
