% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{calcf}
\alias{calcf}
\title{Calculate the true anomaly}
\usage{
calcf(E, e)
}
\value{
true anomaly f (theta)
}
\description{
Calculate the true anomaly f (theta) from the eccentricity
anomaly true anomaly = angle between the point and perihelion
based on Grossman, N.: Orbits under the Inverse Square Law, in
The Sheer Joy of Celestial Mechanics, edited by N. Grossman, pp. 41–61,
Birkhäuser Boston, Boston, MA., 1996.
}
\author{
Thomas Laepple
}
