% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{DailyInsolationParam}
\alias{DailyInsolationParam}
\title{Computes daily average insolation as a function of day and latitude
 given the orbital parameters}
\usage{
DailyInsolationParam(
  lat,
  day,
  ecc,
  obliquity,
  long_perh,
  day_type = 1,
  alpha = 0,
  T.alpha = 80
)
}
\arguments{
\item{lat}{Latitude in degrees (-90 to 90).}

\item{day}{Indicator of time of year; calendar day by default.}

\item{ecc}{eccentricity}

\item{obliquity}{obliquity}

\item{long_perh}{longitude of perihelion}

\item{day_type}{Convention for specifying time of year (+/- 1,2) [optional], see details}

\item{alpha}{parameter for custom calendar,see details}

\item{T.alpha}{parameter for custom calendar,see details}
}
\value{
list(Fsw=Fsw, ecc=ecc, obliquity=obliquity, long_perh=long_perh, lambda=lambda/2/pi*360)
Fsw = Daily average solar radiation in W/m^2.
  ... and all orbital parameters
}
\description{
Insolation calculation, converted and adapted from the Eisenmann and
 Huybers,2006 Matlab Code, based on Berger 1991
}
\section{day type details}{


day_type=1 (default): day input is calendar day (1-365.24), where day 1 is
January first. The calendar is referenced to the vernal equinox which always
occurs at day 80.

day_type=2: day input is solar longitude (0-360 degrees). Solar longitude is
the angle of the Earth's orbit measured from spring equinox (21 March). Note
that calendar days and solar longitude are not linearly related because, by
Kepler's Second Law, Earth's angular velocity varies according to its
distance from the sun.

day_type=3: Custom calendar that alpha (angle since equinox) is reached at
day T.alpha this uses an iterative solution of Keplers equation and is
therefore much slower than the day_type=1 which uses a trigeometric
approximation of Kepler's equation.
}

\author{
Thomas Laepple
}
