% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateTransferFunctionInsolation.R
\name{EstimateTransferfunctionInsolation}
\alias{EstimateTransferfunctionInsolation}
\title{Estimate linear or polynomial transfer function between insolation and
  a climate parameter}
\usage{
EstimateTransferfunctionInsolation(climate, latitude, polynom.order = 3,
  clab = "", bPlot = FALSE, b3plot = FALSE, main = "")
}
\arguments{
\item{latitude}{Latitude (deg North)}

\item{polynom.order}{order of the polynom, 2 or 3}

\item{clab}{name and unit = ylab of climate parameter for plotting}

\item{bPlot}{plot diagnostics (if TRUE)}

\item{b3plot}{plot more diagnostics (if TRUE)}

\item{main}{Title of the plot}
}
\value{
list(coeff, coeff.lin, rsq = c(rsq1, rsq2),lag)
coeff coefficient of polynomial,
coeff.lin = coefficient of linear fit,
rsq of linear and of polynomial fit,
timelag in days
}
\description{
Estimate linear or polynomial transfer function between insolation and
  a climate parameter
}
\examples{
clim <- SelSpace3D(sat.ncep.clim, lat1 = 40, lon1 = 200)

result <- EstimateTransferfunctionInsolation(clim, 70, bPlot=TRUE, b3plot=TRUE)
}
\author{
Thomas Laepple
}
