% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{f.from.t}
\alias{f.from.t}
\title{true angle from days}
\usage{
f.from.t(ecc, t)
}
\arguments{
\item{ecc}{eccentricity}

\item{t}{days from perihelion}
}
\value{
true angle from perihelion (via the mean angle M)
}
\description{
true angle from days
}
\references{
Grossman, N.: Orbits under the Inverse Square Law, in
The Sheer Joy of Celestial Mechanics, edited by N. Grossman, pp. 41–61,
Birkhäuser Boston, Boston, MA., 1996.
}
\author{
Thomas Laepple
}
