% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{glob.grid}
\alias{glob.grid}
\title{Global grid of seasonal cycle amplitudes.}
\format{
A data frame with 10512 rows and 10 variables:
\describe{
  \item{\code{lat}}{double COLUMN_DESCRIPTION}
  \item{\code{lon}}{double COLUMN_DESCRIPTION}
  \item{\code{max.kyear}}{double COLUMN_DESCRIPTION}
  \item{\code{max.amp}}{double COLUMN_DESCRIPTION}
  \item{\code{min.kyear}}{double COLUMN_DESCRIPTION}
  \item{\code{min.amp}}{double COLUMN_DESCRIPTION}
  \item{\code{mean.amp}}{double COLUMN_DESCRIPTION}
  \item{\code{amp.diff}}{double COLUMN_DESCRIPTION}
  \item{\code{sig.sq_c}}{double COLUMN_DESCRIPTION}
  \item{\code{sig.sq_a}}{double COLUMN_DESCRIPTION}
}
}
\usage{
glob.grid
}
\description{
Global grid of seasonal cycle amplitudes for the past 25 ka calculated using
GetSeasCyclePars and transfer functions from EstimateTransferfunctionInsolation usign
sat.ncep.clim
}
\details{
DETAILS
}
\keyword{datasets}
