% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{glob.tf.sat.ncep.clim}
\alias{glob.tf.sat.ncep.clim}
\title{Global grid of insolation to surface temperature transfer functions.}
\format{
A list
}
\usage{
glob.tf.sat.ncep.clim
}
\description{
Global grid of insolation to surface temperature transfer functions.
Calculated using EstimateTransferfunctionInsolation and sat.ncep.clim
}
\details{
DETAILS
}
\keyword{datasets}
