% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{orbital_parameters}
\alias{orbital_parameters}
\title{Load the orbital parameters and return them for a given year}
\usage{
orbital_parameters(kyear, FILEDATA = "e:/data/insolation/data/ins_data.txt")
}
\arguments{
\item{kyear}{time in 1000 years before present (0..5000)}

\item{FILEDATA}{Filename}
}
\value{
list(ecc=ecc,epsilon=epsilon,omega=omega)
}
\description{
Load the orbital parameters and return them for a given year
}
\details{
Load orbital parameters (given each kyr for 0-5Mya)
Load the matrix contains data from Berger and Loutre (1991),
  downloaded as ORBIT91 from ncdc.noaa.gov
}
\author{
Thomas Laepple
}
