% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{orbital_parameters_fast}
\alias{orbital_parameters_fast}
\title{return orbital parameters for a given year using preloaded 100yr resolution values}
\usage{
orbital_parameters_fast(kyear)
}
\arguments{
\item{kyear}{time in 1000 years before present (0..5000)}
}
\value{
list(ecc=ecc,epsilon=epsilon,omega=omega)
}
\description{
return orbital parameters for a given year using preloaded 100yr resolution values
}
\author{
Thomas Laepple
}
