% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-seasonal-cycle-parameters.R
\name{pTs}
\alias{pTs}
\title{Create a pTs object}
\source{
Function copied from "proxytype.R" in paleolibary/src/
}
\usage{
pTs(data, time, lat = 0, lon = 0, name = "", history = "", date = TRUE)
}
\arguments{
\item{data}{a vector, matrix or ts object}

\item{time}{vector}

\item{lat}{vector of length 1}

\item{lon}{vector of length 1}

\item{name}{character string, name of the pTs object to be generated}

\item{history}{character string to append to the history attribute}

\item{date}{logical, whether or not to append the current date to the history attribute}
}
\value{
pTs object
}
\description{
pTs objects are enhanced time-series \code{\link[stats]{ts}} objects. \code{pTs()} adds attributes such as longitude and latitude to a time series
  vector/time series vectors (having the same time basis) and assigns the
  class "pTs" to the resulting object
}
\author{
Thomas Laepple
}
\keyword{internal}
