% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateYearFromInsolation.R
\name{SimulateYearFromInsolation}
\alias{SimulateYearFromInsolation}
\title{Simulate a year from a parametric relationship to insolation}
\usage{
SimulateYearFromInsolation(kyear = 0, transfer, latitude,
  bPolynomial = FALSE, anomalyRef = NULL)
}
\arguments{
\item{kyear}{point in time (kyr BP)}

\item{transfer}{transfer function}

\item{latitude}{latitude of the insolation calculation}

\item{bPolynomial}{(TRUE) use polynomial, FALSE = linear}

\item{anomalyRef}{NULL for absolute mode or vector of 365 anomaly correction values}
}
\value{
365 predicted values for each day of the year
}
\description{
Simulate a year from a parametric relationship to insolation
}
\examples{
clim <- ecustools::SelSpace3D(sat.ncep.clim, lat1 = 40, lon1 = 200)
transfer <- EstimateTransferfunctionInsolation(clim, 40, bPlot = FALSE, b3plot = FALSE)
plot(1:365, SimulateYearFromInsolation(0, transfer, 40), type = "l",
     main = "predicted at 0 and 10kyr BP", xlab = "day of the year")
lines(1:365, SimulateYearFromInsolation(10, transfer, 40), col = "red")
}
\author{
Thomas Laepple
}
