% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{t.from.f}
\alias{t.from.f}
\title{days to true angle}
\usage{
\method{t}{from.f}(ecc, f)
}
\arguments{
\item{ecc}{eccentricity}

\item{f}{true angle from perihelion}
}
\value{
days from perihelion, calculated using the mean angle M
}
\description{
days to true angle
}
\references{
Grossman, N.: Orbits under the Inverse Square Law, in
The Sheer Joy of Celestial Mechanics, edited by N. Grossman, pp. 41–61,
Birkhäuser Boston, Boston, MA., 1996.
}
\author{
Thomas Laepple
}
