% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateYearFromInsolation.R
\name{TLag}
\alias{TLag}
\title{timelag with rolling over at 365 days}
\usage{
TLag(data, ilag)
}
\arguments{
\item{data}{vector of values at 365 days}

\item{ilag}{lag in days}
}
\value{
lagged values
}
\description{
timelag with rolling over at 365 days
}
\author{
Thomas Laepple
}
