% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insolation.R
\name{unwrap}
\alias{unwrap}
\title{Correct the radian phase angles}
\usage{
unwrap(p)
}
\arguments{
\item{p}{radian phase angles array}
}
\value{
unwrapped angles
}
\description{
Correct the radian phase angles
}
\details{
Q = unwrap(P) corrects the radian phase angles in array P by adding
multiples of \eqn{+- 2\pi} when absolute jumps between consecutive array
elements are greater than pi radians.
based on http://ccrma.stanford.edu/~jos/sasp/Matlab_listing_unwrap_m.html
}
\author{
Thomas Laepple
}
