% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-seasonal-cycle-parameters.R
\name{WhichTF}
\alias{WhichTF}
\title{Helper function for GetSeasCyclePars}
\usage{
WhichTF(lon, lat)
}
\arguments{
\item{lat}{}
}
\value{

}
\description{
Helper function for GetSeasCyclePars
}
\examples{
WhichTF(lon = 20, lat = 2.5)
}
